Make the schema fully immutable and happy.
Previously we were instantiating each rule as it was added to the schema. This
was fine for basic rules, but some rules became more advanced and actually
modified other rules when they were compiled. This changes the schema so
that rules aren't actually instantiated until we start editing the schema,
and when they are they're stored separately and that chain of rules is
never modified. This, I think, should serve to make the schemas
totally immutable, without any runtime cost.